+++ src/HttpRepository/AbstractMicroserviceHttpRepository.php
@@ -62,14 +62,13 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
     abstract protected function getResourceDto(): string;
 
     /**
-     * @psalm-param array<array-key, mixed> $additionalQueryParams
+     * @param array<string, mixed> $additionalQueryParams
      *
      * @throws ExceptionInterface
      */
     public function fetchOneByIri(string $iri, array $additionalQueryParams = []): ?ApiResourceDtoInterface
     {
         try {
-            /* @var ApiResourceDtoInterface|null */
             return $this->serializer->deserialize(
                 $this->request('GET', $this->buildUri($iri, $additionalQueryParams))->getContent(),
                 $this->getResourceDto(),
@@ -87,7 +86,7 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
     /**
      * @deprecated since version 0.3.0, will be removed in 1.0. Use {@see \Mtarld\ApiPlatformMsBundle\HttpRepository\AbstractMicroserviceHttpRepository::fetchOneByIri()} instead.
      *
-     * @psalm-param array<array-key, mixed> $additionalQueryParams
+     * @param array<string, mixed> $additionalQueryParams
      *
      * @throws ExceptionInterface
      */
@@ -97,8 +96,8 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
     }
 
     /**
-     * @psalm-param array<array-key, mixed> $criteria
-     * @psalm-param array<array-key, mixed> $additionalQueryParams
+     * @param array<string, mixed> $criteria
+     * @param array<string, mixed> $additionalQueryParams
      *
      * @throws ExceptionInterface
      */
@@ -113,8 +112,8 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
     /**
      * @deprecated since version 0.3.0, will be removed in 1.0. Use {@see \Mtarld\ApiPlatformMsBundle\HttpRepository\AbstractMicroserviceHttpRepository::fetchOneBy()} instead.
      *
-     * @psalm-param mixed $value
-     * @psalm-param array<array-key, mixed> $additionalQueryParams
+     * @param mixed                $value
+     * @param array<string, mixed> $additionalQueryParams
      *
      * @throws ExceptionInterface
      */
@@ -124,9 +123,10 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
     }
 
     /**
-     * @psalm-param array<array-key, mixed> $criteria
-     * @psalm-param array<array-key, mixed> $additionalQueryParams
-     * @psalm-return Collection<ApiResourceDtoInterface>
+     * @param array<string, mixed> $criteria
+     * @param array<string, mixed> $additionalQueryParams
+     *
+     * @return Collection<ApiResourceDtoInterface>
      *
      * @throws ExceptionInterface
      */
@@ -138,9 +138,10 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
     /**
      * @deprecated since version 0.3.0, will be removed in 1.0. Use {@see \Mtarld\ApiPlatformMsBundle\HttpRepository\AbstractMicroserviceHttpRepository::fetchBy()} instead.
      *
-     * @param array<mixed> $values
-     * @psalm-param array<array-key, mixed> $additionalQueryParams
-     * @psalm-return Collection<ApiResourceDtoInterface>
+     * @param array<mixed>         $values
+     * @param array<string, mixed> $additionalQueryParams
+     *
+     * @return Collection<ApiResourceDtoInterface>
      *
      * @throws ExceptionInterface
      */
@@ -150,8 +151,9 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
     }
 
     /**
-     * @psalm-param array<array-key, mixed> $additionalQueryParams
-     * @psalm-return Collection<ApiResourceDtoInterface>
+     * @param array<string, mixed> $additionalQueryParams
+     *
+     * @return Collection<ApiResourceDtoInterface>
      *
      * @throws ExceptionInterface
      */
@@ -163,8 +165,9 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
     /**
      * @deprecated since version 0.3.0, will be removed in 1.0. Use {@see \Mtarld\ApiPlatformMsBundle\HttpRepository\AbstractMicroserviceHttpRepository::fetchAll()} instead.
      *
-     * @psalm-param array<array-key, mixed> $additionalQueryParams
-     * @psalm-return Collection<ApiResourceDtoInterface>
+     * @param array<string, mixed> $additionalQueryParams
+     *
+     * @return Collection<ApiResourceDtoInterface>
      *
      * @throws ExceptionInterface
      */
@@ -174,7 +177,7 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
     }
 
     /**
-     * @psalm-param array<array-key, mixed> $additionalQueryParams
+     * @param array<string, mixed> $additionalQueryParams
      *
      * @throws ResourceValidationException
      * @throws ExceptionInterface
@@ -184,7 +187,6 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
         try {
             $response = $this->request('POST', $this->buildUri($this->getResourceEndpoint(), $additionalQueryParams), $resource, null, 'json');
 
-            /* @var ApiResourceDtoInterface */
             return $this->serializer->deserialize($response->getContent(), $this->getResourceDto(), $this->getMicroservice()->getFormat());
         } catch (ClientExceptionInterface $e) {
             if ((400 === $e->getCode()) && null !== $violations = $this->createConstraintViolationListFromResponse($e->getResponse())) {
@@ -198,7 +200,8 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
     /**
      * Update a resource using PUT verb.
      *
-     * @psalm-param array<array-key, mixed> $additionalQueryParams
+     * @param array<string, mixed> $additionalQueryParams
+
      *
      * @throws ResourceValidationException
      * @throws ExceptionInterface
@@ -213,7 +216,6 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
         try {
             $response = $this->request('PUT', $this->buildUri($iri, $additionalQueryParams), $resource, null, 'json');
 
-            /* @var ApiResourceDtoInterface */
             return $this->serializer->deserialize($response->getContent(), $this->getResourceDto(), $this->getMicroservice()->getFormat());
         } catch (ClientExceptionInterface $e) {
             if ((400 === $e->getCode()) && null !== $violations = $this->createConstraintViolationListFromResponse($e->getResponse())) {
@@ -222,7 +224,6 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
 
             throw $e;
         }
-    }
 
     /**
      * Partially update a resource using PATCH verb.
@@ -254,7 +255,7 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
     }
 
     /**
-     * @psalm-param array<array-key, mixed> $additionalQueryParams
+     * @param array<string, mixed> $additionalQueryParams
      *
      * @throws ResourceValidationException
      * @throws ExceptionInterface
@@ -287,8 +288,9 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
     }
 
     /**
-     * @psalm-param array<array-key, mixed> $queryParams
-     * @psalm-return Collection<ApiResourceDtoInterface>
+     * @param array<string, mixed> $queryParams
+     *
+     * @return Collection<ApiResourceDtoInterface>
      *
      * @throws ExceptionInterface
      */
@@ -296,7 +298,7 @@ abstract class AbstractMicroserviceHttpRepository implements ReplaceableHttpClie
     {
         $response = $this->request('GET', $this->buildUri($this->getResourceEndpoint(), $queryParams));
 
-        /* @var Collection $collection */
+        /** @var Collection<ApiResourceDtoInterface> $collection */
         $collection = $this->serializer->deserialize(
             $response->getContent(),
             Collection::class.'<'.$this->getResourceDto().'>',